Add run-winfstest.ps1
We'll add another help that's supposed to run the tests. Unlinke
the existing cmd script, it's expecting python to be part of %PATH%
instead of attempting to locate it using registry keys that can
change across Python versions.

While at it, we're replacing the batch script using Powershell,
which is arguably easier to read.